+++ tests/test_flip.py
@@ -13,138 +13,6 @@ import json
 import unittest
 import yaml
 
-class ReplaceJoinTestCase(unittest.TestCase):
-    """
-    Check that joins get replaced with Subs
-    """
-
-    def test_basic_case(self):
-        """
-        As simple as it gets
-        """
-
-        source = {
-            "Fn::Join": [
-                " ",
-                ["The", "cake", "is", "a", "lie"],
-            ],
-        }
-
-        expected = "The cake is a lie"
-
-        actual = cfn_flip.clean(source)
-
-        self.assertEqual(expected, actual)
-
-    def test_ref(self):
-        """
-        Refs should be replaced by ${value}
-        """
-
-        source = {
-            "Fn::Join": [
-                " ",
-                ["The", {"Ref": "Cake"}, "is", "a", "lie"],
-            ],
-        }
-
-        expected = {
-            "Fn::Sub": "The ${Cake} is a lie",
-        }
-
-        actual = cfn_flip.clean(source)
-
-        self.assertEqual(expected, actual)
-
-    def test_get_att(self):
-        """
-        Base64 etc should be replaced by parameters to Sub
-        """
-
-        source = {
-            "Fn::Join": [
-                " ",
-                ["The", {"Fn::GetAtt": ["Cake", "Hole"]}, "is", "a", "lie"],
-            ],
-        }
-
-        expected = {
-            "Fn::Sub": "The ${Cake.Hole} is a lie",
-        }
-
-        actual = cfn_flip.clean(source)
-
-        self.assertEqual(expected, actual)
-
-    def test_others(self):
-        """
-        GetAtt should be replaced by ${Thing.Property}
-        """
-
-        source = {
-            "Fn::Join": [
-                " ",
-                ["The", {"Fn::Base64": "Notreallybase64"}, "is", "a", "lie"],
-            ],
-        }
-
-        expected = {
-            "Fn::Sub": [
-                "The ${Param1} is a lie",
-                {
-                    "Param1": {
-                        "Fn::Base64": "Notreallybase64",
-                    },
-                },
-            ],
-        }
-
-        actual = cfn_flip.clean(source)
-
-        self.assertEqual(expected, actual)
-
-    def test_in_array(self):
-        """
-        Converting Join to Sub should still work when the join is part of a larger array
-        """
-
-        source = {
-            "things": [
-                "Just a string",
-                {
-                    "Fn::Join": [
-                        " ",
-                        ["The", {"Fn::Base64": "Notreallybase64"}, "is", "a", "lie"],
-                    ],
-                },
-                {
-                    "Another": "thing",
-                },
-            ],
-        }
-
-        expected = {
-            "things": [
-                "Just a string",
-                {
-                    "Fn::Sub": [
-                        "The ${Param1} is a lie",
-                        {
-                            "Param1": {
-                                "Fn::Base64": "Notreallybase64",
-                            },
-                        },
-                    ],
-                },
-                {
-                    "Another": "thing",
-                },
-            ],
-        }
-
-        actual = cfn_flip.clean(source)
-
-        self.assertEqual(expected, actual)
 
 class CfnFlipTestCase(unittest.TestCase):
     def setUp(self):
@@ -341,6 +209,3 @@ class CfnFlipTestCase(unittest.TestCase):
 
         actual = cfn_flip.to_json(source, clean_up=True)
         self.assertEqual(expected, json.loads(actual))
-
-if __name__ == "__main__":
-    unittest.main()
